+++ example/lib/main.dart
@@ -71,9 +71,7 @@ class _HomePageState extends State<HomePage> {
         kDebugMode ? debugPrint('ERROR WAITING FOR READY: $e') : null;
       }
     } on PlatformException {
-      kDebugMode
-          ? debugPrint("Error: An error occurred while starting the recording!")
-          : null;
+      kDebugMode ? debugPrint("Error: An error occurred while starting the recording!") : null;
     }
   }
 
@@ -84,9 +82,23 @@ class _HomePageState extends State<HomePage> {
         _response = stopResponse;
       });
     } on PlatformException {
-      kDebugMode
-          ? debugPrint("Error: An error occurred while stopping recording.")
-          : null;
+      kDebugMode ? debugPrint("Error: An error occurred while stopping recording.") : null;
+    }
+  }
+
+  Future<void> pauseRecord() async {
+    try {
+      await screenRecorder?.pauseRecord();
+    } on PlatformException {
+      kDebugMode ? debugPrint("Error: An error occurred while pause recording.") : null;
+    }
+  }
+
+  Future<void> resumeRecord() async {
+    try {
+      await screenRecorder?.resumeRecord();
+    } on PlatformException {
+      kDebugMode ? debugPrint("Error: An error occurred while resume recording.") : null;
     }
   }
 
@@ -108,12 +120,10 @@ class _HomePageState extends State<HomePage> {
             Text("Video Hash: ${_response?['videohash']}"),
             Text("Start Date: ${(_response?['startdate']).toString()}"),
             Text("End Date: ${(_response?['enddate']).toString()}"),
-            ElevatedButton(
-                onPressed: () => startRecord(fileName: "eren"),
-                child: const Text('START RECORD')),
-            ElevatedButton(
-                onPressed: () => stopRecord(),
-                child: const Text('STOP RECORD')),
+            ElevatedButton(onPressed: () => startRecord(fileName: "eren"), child: const Text('START RECORD')),
+            ElevatedButton(onPressed: () => resumeRecord(), child: const Text('RESUME RECORD')),
+            ElevatedButton(onPressed: () => pauseRecord(), child: const Text('PAUSE RECORD')),
+            ElevatedButton(onPressed: () => stopRecord(), child: const Text('STOP RECORD')),
           ],
         ),
       ),