+++ lib/src/ed_screen_recorder/ed_screen_recorder_plugin.dart
@@ -87,4 +87,14 @@ class EdScreenRecorder {
     }
     return formatResponse.toJson();
   }
+
+  Future<void> pauseRecord() async {
+    await _channel.invokeMethod('pauseRecordScreen');
+    print("Pause Record");
+  }
+
+  Future<void> resumeRecord() async {
+    await _channel.invokeMethod('resumeRecordScreen');
+    print("Resume Record");
+  }
 }